requirements.txt
aiohttp==3.5.4
aiohttp-toolbox==0.3.3
brotli==1.0.7
cchardet==2.1.4
pillow==5.3.0
pydantic==0.20
raven==6.10.0
raven-aiohttp==0.7.0
uvloop==0.11.3
